Fixing bugs with login

+++ my_website/main/views.py
@@ -2,20 +2,45 @@ from django.shortcuts import render, redirect
 from django.http import HttpResponse, HttpResponseRedirect
 from django.contrib import messages
 from django.urls import reverse
+from django.views import View
+from django.views.decorators.csrf import csrf_protect
+
 from .models import ToDoList
 from .forms import CreateNewList
 
 # Create your views here.
 
-def delete(request, id):
-    tdTable = ToDoList.objects.get(id=id)
-    messages.info(request, f"Successfully deleted {tdTable.name}")
-    tdTable.delete()
-    return redirect(reverse('view'))
+class ToDoListView(View):
+    template_name = "main/view.html"
 
-def home(request):
-    return render(request, "main/home.html", {})
+    def get(self, request):
+        form = CreateNewList()
+        return render(request, self.template_name, {"form": form})
+    
+    def post(self, request):
+        form = CreateNewList(request.POST)
+        delete_id = request.POST.get("delete_id")
+        
+        if form.is_valid():
+            name = form.cleaned_data["name"]
+            todolist = ToDoList(name=name)
+            todolist.save()
+            request.user.todolist.add(todolist)
 
+            return redirect("/%i" %todolist.id)
+        elif delete_id:
+            tdTable = ToDoList.objects.get(id=delete_id)
+            messages.info(request, f"Successfully deleted {tdTable.name}")
+            tdTable.delete()
+
+        return redirect('view')
+
+class HomeView(View):
+    template_name = "main/home.html"
+
+    def get(self, request):
+        return render(request, self.template_name)
+    
 def index(request, id):
     ls = ToDoList.objects.get(id=id)
 
@@ -45,21 +70,3 @@ def index(request, id):
 
     return render(request, "main/list.html", {"ls":ls})
 
-def todolist(request):
-    ls = request.user.todolist.all().count()
-    return render(request, "main/todolist.html", {"ls":ls})
-
-def view(request):
-    if request.method == "POST":
-        form = CreateNewList(request.POST)
-
-        if form.is_valid():
-            n = form.cleaned_data["name"]
-            t = ToDoList(name=n)
-            t.save()
-            request.user.todolist.add(t)
-        
-        return HttpResponseRedirect("/%i" %t.id)
-    else:
-        form = CreateNewList()
-    return render(request, "main/view.html", {"form":form})
